SP-5025 BASIC Res Parascan
Res Parascan. Find two resistors from a list of values stored in the DATA statements that can be used in parallel to make a non-standard resistance value within a tolerance of x%. Based on the E24 Parascan program and the limited range of resistors I generally have to hand - so slightly more practical.