SP-5025 BASIC - Duck Shoot
A BASIC game, credited in the listing to Sourthern Software, 1980. Probably typed into my MZ-80K from a magazine like Computer and Video Games in 1982/82 by my brother. There are a few obvious bugs in the code, but not enough to get enjoyment from it if shooting simulated wildlife is the way you get your kicks.